make sure lang is loaded before we start the upgrade
At the end of the upgrade, mdkonline-upgrade-helper invokes some ugtk2
code which does a require lang; which in turn tries to load utf8.pm (a
module part of the core perl distribution). But if perl was upgraded to
a newer version as part of the upgrade, utf8.pm might have moved to
location different from where the running perl expects to find it (eg
mdkonline-upgrade-helper was started by perl 5.8.0 which looks for its
files in a directory containing 5.8.0 but in the mean time perl 5.10.1
was installed so utf8.pm is in a directory containing 5.10.1 in its
name) (#55090).

git-svn-id: svn+ssh://svn.mandriva.com/svn/soft/mdkonline/trunk@263280 99302b65-d5f7-0310-b3dd-f8cd6f4e3d94